pearls: missing numbers test passing

+++ pearls/pearls/missing_number.h
@@ -4,8 +4,8 @@
 #include <vector>
 
 namespace ch2 {
-//    finds a missing number in numbers (range is 1, max(numbers))
-    int get_missing_number(std::vector<int> &numbers);
+//    finds a missing number in numbers (range is [0, max(numbers)])
+    ssize_t get_missing_number(std::vector<size_t> &numbers);
 }
 
 #endif //PEARLS_MISSING_NUMBER_H
