+++ pearls/tests/ch2/test_missing_number.cpp
@@ -3,6 +3,6 @@
 
 
 TEST(MissingNumber, ItFindsMissingNumber) {
-    std::vector<int> numbers = {3, 4, 1, 5};
+    std::vector<size_t> numbers = {3, 4, 1, 5};
     EXPECT_EQ(ch2::get_missing_number(numbers), 2);
 }
